refactor

+++ src/main/java/com/example/mag/aktuelapi/service/CategoryService.java
@@ -1,10 +1,15 @@
 package com.example.mag.aktuelapi.service;
 
-import com.example.mag.aktuelapi.dto.category.CategoryDto;
+import com.example.mag.aktuelapi.dto.category.CategoryDtoRequest;
+import com.example.mag.aktuelapi.dto.category.CategoryDtoResponse;
+import com.example.mag.aktuelapi.dto.mark.MarkDtoResponse;
 import com.example.mag.aktuelapi.model.Category;
 import com.example.mag.aktuelapi.repository.CategoryRepository;
 import org.springframework.stereotype.Service;
 
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Base64;
 import java.util.List;
 import java.util.Optional;
 
@@ -12,39 +17,56 @@ import java.util.Optional;
 public class CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final MarkService markService;
 
-    public CategoryService(CategoryRepository categoryRepository) {
+    public CategoryService(CategoryRepository categoryRepository, MarkService markService) {
         this.categoryRepository = categoryRepository;
+        this.markService = markService;
     }
 
 
     public Optional<Category> getCategoryId(Long id) {
-       return categoryRepository.findById(id);
+        return categoryRepository.findById(id);
     }
 
-    public List<Category> findAll() {
-        return categoryRepository.findAll();
+    public List<CategoryDtoResponse> getAll() {
+        List<CategoryDtoResponse> categoryDtoResponses = new ArrayList<>();
+        List<Category> categorieList = categoryRepository.findAll();
+        categorieList.forEach(category -> {
+            CategoryDtoResponse categoryDtoResponse = new CategoryDtoResponse();
+            categoryDtoResponse.setId(category.getId());
+            categoryDtoResponse.setName(category.getName());
+            categoryDtoResponse.setImage(category.getImage()==null?"": Base64.getEncoder().encodeToString(category.getImage()));
+            categoryDtoResponses.add(categoryDtoResponse);
+        });
+        return categoryDtoResponses;
     }
 
-    public Category create(CategoryDto categoryDto) {
+    public Category create(CategoryDtoRequest dtoRequest) throws IOException {
         Category category = new Category();
-        category.setName(categoryDto.getName());
-        category.setLink(categoryDto.getLink());
+        category.setName(dtoRequest.getName());
+        category.setImage(dtoRequest.getImage().getBytes());
         categoryRepository.save(category);
         return category;
     }
 
-    public Category update(Long id, CategoryDto dto) {
-        return categoryRepository.findById(id).map(product -> {
-            product.setName(dto.getName());
-            product.setLink(dto.getLink());
-            return categoryRepository.save(product);
-        }).orElseThrow(() -> new RuntimeException("Mark not found"));
+    public Category update(Long id, CategoryDtoRequest categoryDtoRequest) throws IOException {
+        Category category = categoryRepository.findById(id).orElseThrow(() -> new RuntimeException("Category not found"));
+        category.setName(categoryDtoRequest.getName());
+        if (categoryDtoRequest.getImage() != null) {
+            category.setImage(categoryDtoRequest.getImage().getBytes());
+        }
+        return categoryRepository.save(category);
     }
 
     public void delete(Long id) {
         if (categoryRepository.existsById(id)) {
-            categoryRepository.deleteById(id);
+            List<MarkDtoResponse> markDtoResponses = markService.getMarkByCategoryId(id);
+            if (markDtoResponses.isEmpty()) {
+                categoryRepository.deleteById(id);
+            } else {
+                throw new RuntimeException("Bu category ait marka degerleri oldugu icin category silemezsiniz");
+            }
         } else {
             throw new RuntimeException("Kayıt bulunamadı");
         }
